refactored show page to only show appropriate open or close button per current state, finished moving logic to Question Form component from Forums
